HDM-422 Move IU behaviors to /app directory
