+++ app/models/concerns/file_set_behavior.rb
@@ -1,0 +1,10 @@
+require 'hydradam/file_set_behavior/has_ffprobe'
+
+module Concerns
+  module FileSetBehavior
+    extend ActiveSupport::Concern
+    include HydraDAM::FileSetBehavior::HasFfprobe
+
+    attr_accessor :quality_level
+  end
+end
