+++ app/models/file_set.rb
@@ -1,8 +1,6 @@
-require 'IU/models/concerns/file_set_behavior'
-
 class FileSet < ActiveFedora::Base
-  include IU::Models::Concerns::FileSetBehavior
-
+  include ::CurationConcerns::FileSetBehavior
+  include Concerns::FileSetBehavior
 
   class << self
     def indexer
