+++ app/models/work.rb
@@ -1,14 +1,11 @@
 # Generated via
 #  `rails generate curation_concerns:work Work`
-require 'IU/models/concerns/work_behavior'
-
-
 class Work < ActiveFedora::Base
 
   FIXITY_TYPE = :md5
   include ::CurationConcerns::WorkBehavior
   include ::CurationConcerns::BasicMetadata
-  include ::IU::Models::Concerns::WorkBehavior
+  include Concerns::WorkBehavior
   # validates :title, presence: { message: 'Your work must have a title.' }
 
   def do_md5_checksum
