--- lib/IU/models/file_set.rb
@@ -1,10 +1,0 @@
-require 'IU/models/concerns/file_set_behavior'
-
-module IU
-  module Models
-    class FileSet < ActiveFedora::Base
-      include CurationConcerns::FileSetBehavior
-      include IU::Models::Concerns::FileSetBehavior
-    end
-  end
-end
