+++ spec/lib/IU/models/file_set_spec.rb
@@ -1,7 +1,7 @@
 require 'rails_helper'
 
 describe FileSet do
-  it 'includes IU::Models::Concerns::FileSetBehavior' do
-    expect(described_class.ancestors).to include IU::Models::Concerns::FileSetBehavior
+  it 'includes Concerns::FileSetBehavior' do
+    expect(described_class.ancestors).to include Concerns::FileSetBehavior
   end
 end
